+++ app/src/organisms/ProtocolSetup/RunSetupCard/LabwareSetup/SecureLabwareModal.tsx
@@ -21,7 +21,7 @@ import {
   SPACING_4,
   SPACING_5,
 } from '@opentrons/components'
-import { Portal } from '../../../App/portal'
+import { Portal } from '../../../../App/portal'
 import secureMagModBracketImage from '../../../assets/images/secure_mag_mod_bracket.svg'
 import secureTCLatchImage from '../../../assets/images/secure_tc_latch.png'
 import { getModuleName } from './utils/getModuleName'
