+++ app/src/organisms/ProtocolSetup/RunSetupCard/LabwareSetup/__tests__/LabwareInfoOverlay.test.tsx
@@ -6,7 +6,7 @@ import {
 } from '@opentrons/shared-data'
 import fixture_tiprack_300_ul from '@opentrons/shared-data/labware/fixtures/2/fixture_tiprack_300_ul.json'
 import { renderWithProviders } from '@opentrons/components/__utils__'
-import { i18n } from '../../../../i18n'
+import { i18n } from '../../../../../i18n'
 import { LabwareInfoOverlay } from '../LabwareInfoOverlay'
 
 jest.mock('@opentrons/shared-data', () => {
