+++ app/src/organisms/ProtocolSetup/RunSetupCard/LabwareSetup/__tests__/LabwareSetupModal.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { fireEvent } from '@testing-library/react'
 import { renderWithProviders } from '@opentrons/components/__utils__'
-import { i18n } from '../../../../i18n'
+import { i18n } from '../../../../../i18n'
 import { LabwareSetupModal } from '../LabwareSetupModal'
 
 const render = (props: React.ComponentProps<typeof LabwareSetupModal>) => {
