+++ app/src/organisms/ProtocolSetup/RunSetupCard/LabwareSetup/utils/getLabwareRenderCoords.ts
