+++ app/src/organisms/ProtocolSetup/RunSetupCard/ModuleSetup/ModuleInfo.tsx
