+++ app/src/organisms/ProtocolSetup/RunSetupCard/index.tsx
@@ -12,10 +12,10 @@ import standardDeckDef from '@opentrons/shared-data/deck/definitions/2/ot2_stand
 import { getProtocolData } from '../../../redux/protocol'
 import { Divider } from '../../../atoms/structure'
 import { CollapsibleStep } from './CollapsibleStep'
-import { LabwareSetup } from '../LabwareSetup'
-import { ModuleSetup } from '../ModuleSetup'
-import { getModuleRenderCoords } from '../LabwareSetup/utils/getModuleRenderCoords'
-import { getLabwareRenderCoords } from '../LabwareSetup/utils/getLabwareRenderCoords'
+import { LabwareSetup } from './LabwareSetup'
+import { ModuleSetup } from './ModuleSetup'
+import { getModuleRenderCoords } from './LabwareSetup/utils/getModuleRenderCoords'
+import { getLabwareRenderCoords } from './LabwareSetup/utils/getLabwareRenderCoords'
 
 import { RobotCalibration } from './RobotCalibration'
 import type { JsonProtocolFile } from '@opentrons/shared-data'
